+++ packages/g-plugin-webgpu-device/src/platform/utils.ts
@@ -52,10 +52,18 @@ export function translateTextureUsage(
 }
 
 export function translateTextureFormat(format: Format): GPUTextureFormat {
-  if (format === Format.U8_RGBA_RT) return 'bgra8unorm';
-  else if (format === Format.U8_RGBA_NORM) return 'rgba8unorm';
+  if (format === Format.U8_R_NORM) return 'r8unorm';
   else if (format === Format.U8_RG_NORM) return 'rg8unorm';
+  else if (format === Format.U8_RGBA_RT) return 'bgra8unorm';
+  else if (format === Format.U8_RGBA_RT_SRGB) return 'bgra8unorm-srgb';
+  else if (format === Format.U8_RGBA_NORM) return 'rgba8unorm';
+  else if (format === Format.U8_RGBA_SRGB) return 'rgba8unorm-srgb';
+  else if (format === Format.S8_R_NORM) return 'r8snorm';
+  else if (format === Format.S8_RG_NORM) return 'rg8snorm';
+  else if (format === Format.S8_RGBA_NORM) return 'rgba8snorm';
   else if (format === Format.U32_R) return 'r32uint';
+  else if (format === Format.F16_RGBA) return 'rgba16float';
+  else if (format === Format.F32_RGBA) return 'rgba32float';
   else if (format === Format.D24) return 'depth24plus';
   else if (format === Format.D24_S8) return 'depth24plus-stencil8';
   else if (format === Format.D32F) return 'depth32float';
@@ -70,15 +78,15 @@ export function translateTextureFormat(format: Format): GPUTextureFormat {
   else if (format === Format.BC4_UNORM) return 'bc4-r-unorm';
   else if (format === Format.BC5_SNORM) return 'bc5-rg-snorm';
   else if (format === Format.BC5_UNORM) return 'bc5-rg-unorm';
-  else throw new Error('whoops');
+  else throw 'whoops';
 }
 
 export function translateTextureDimension(
   dimension: TextureDimension,
 ): GPUTextureDimension {
   if (dimension === TextureDimension.n2D) return '2d';
-  else if (dimension === TextureDimension.Cube) return '3d';
-  else if (dimension === TextureDimension.n2DArray) return '3d';
+  else if (dimension === TextureDimension.Cube) return '2d';
+  else if (dimension === TextureDimension.n2DArray) return '2d';
   else if (dimension === TextureDimension.n3D) return '3d';
   else throw new Error('whoops');
 }