+++ packages/g-math/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @antv/g-math
 
+## 2.0.1
+
+### Patch Changes
+
+-   b0dd4788: Remove this syntax in @antv/g-math since it's already an ESM now
+
 ## 2.0.0
 
 ### Major Changes