+++ packages/g-plugin-3d/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @antv/g-plugin-3d
 
+## 1.9.1
+
+### Patch Changes
+
+-   b0dd4788: Remove this syntax in @antv/g-math since it's already an ESM now
+-   Updated dependencies [b0dd4788]
+    -   @antv/g-plugin-device-renderer@1.9.1
+    -   @antv/g-shader-components@1.8.1
+    -   @antv/g-lite@1.2.1
+
 ## 1.9.0
 
 ### Minor Changes