+++ packages/g-plugin-webgl-device/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @antv/g-plugin-webgl-device
 
+## 1.9.1
+
+### Patch Changes
+
+-   b0dd4788: Remove this syntax in @antv/g-math since it's already an ESM now
+-   Updated dependencies [b0dd4788]
+    -   @antv/g-plugin-device-renderer@1.9.1
+    -   @antv/g-lite@1.2.1
+
 ## 1.9.0
 
 ### Minor Changes